requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.1.0

# Additional dependencies for weather functionality
requests>=2.31.0
python-dotenv>=1.0.0

# Development tools
ruff>=0.1.0
mypy>=1.7.0

